Move GroupedObservable to rx.observables
